countryController-validation

+++ app/Traits/ApiCrudTrait.php
@@ -13,34 +13,31 @@
 	use Illuminate\Http\Request;
 	use Illuminate\Support\Facades\App;
 	use Illuminate\Support\Facades\Validator;
+    use Illuminate\Validation\ValidationException;
 
-	trait ApiCrudTrait
+    trait ApiCrudTrait
 
 	{
 
 		abstract function model();
-
 		abstract function validationRules($resource_id =0);
 		abstract function columns();
 
 
-		/*
-			public function index(array $columns = ['*'], array $relations = [])
-			{
-				return $this->model()::all();
-			}
-			*/
 
-			public function index(array $columns = ['*'], array $relations = []): Collection
-			{
-				return $this->model()::with($relations)->select($this->columns())->get();
-			}
-
+        public function index( ): Collection
+        {
+            return $this->model()::all();
+        }
 			public function create(Request $request)
 			{
-				Validator::make($request->all(), $this->validationRules())->validate();
+			  $validator = 	Validator::make($request->all(), $this->validationRules()) ;
+                if ($validator->fails()) {
+                    return new ValidationException($validator);
+                }
 
 				return $this->model()::create($request->all());
+
 			}
 
 			public function show($resource_id)
@@ -50,18 +47,29 @@
 
 			public function update(Request $request, $resource_id)
 			{
-				$resource = $this->model()::findOrFail($resource_id);
+				$resource = $this->model()::find($resource_id);
 
-				Validator::make($request->all(), $this->validationRules($resource_id))->validate();
-
-				return $resource->update($request->all());
+                if(empty($resource)){
+                    return response('not-found',404);
+                }else{
+                $validator = 	Validator::make($request->all(), $this->validationRules($resource_id)) ;
+                if ($validator->fails()) {
+                    return new ValidationException($validator);
+                }
+				 $resource->update($request->all());
+                    return $resource;
+                }
 			}
 
 			public function delete($resource_id)
 			{
-				$resource = $this->model()::findOrFail($resource_id);
-
-				return $resource->delete();
+				$resource = $this->model()::find($resource_id);
+                if(empty($resource)){
+                    return response('not-found',404);
+                }else {
+                     $resource->delete();
+                    return response('deleted',200);
+                }
 			}
 
 	}
